+++ app/src/main/java/com/kyhsgeekcode/disassembler/ui/components/TreeView.kt
@@ -1,4 +1,4 @@
-package com.kyhsgeekcode.disassembler.ui
+package com.kyhsgeekcode.disassembler.ui.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
