+++ app/src/main/java/com/kyhsgeekcode/disassembler/ui/tabs/BinaryDisasmTab.kt
@@ -19,8 +19,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.kyhsgeekcode.disassembler.*
-import com.kyhsgeekcode.disassembler.ui.InfiniteList
+
 import com.kyhsgeekcode.disassembler.ui.components.CellText
+import com.kyhsgeekcode.disassembler.ui.components.InfiniteList
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
