+++ app/src/main/java/com/kyhsgeekcode/disassembler/ui/tabs/BinaryOverviewTab.kt
@@ -6,7 +6,6 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Button
 import androidx.compose.material.Text
-import androidx.compose.material.TextField
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -14,8 +13,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.kyhsgeekcode.disassembler.AbstractFile
 import com.kyhsgeekcode.disassembler.R
-import com.kyhsgeekcode.disassembler.ui.HexTextField
-import com.kyhsgeekcode.disassembler.ui.Spinner
+import com.kyhsgeekcode.disassembler.ui.components.HexTextField
+import com.kyhsgeekcode.disassembler.ui.components.Spinner
 import nl.lxtreme.binutils.elf.MachineType
 import java.util.*
